웨이브 구현중, 추후 해야할거(VFX 미싱 나는거 고치기+ 웨이브 여러 번 나오게끔 코드 엎기, 배경 추가)

+++ Assets/Scripts/Scriptable/DefaultDatable.cs
@@ -5,7 +5,7 @@ public class DefaultDatable : ScriptableObject
 {
     public int id;
     public int Hp;
-    public int spd;
+    public float spd;
     public int atk;
     public float atkspd;
 }
